add documentation and refactor classes
