+++ src/LangVisitor.py
@@ -1,7 +1,7 @@
 from arpeggio import PTNodeVisitor
 from CodeEvaluator import *
 from ExpressionEvaluator import *
-from LangEvaluator import *
+from LangContext import *
 from VariableEvaluator import *
 from FunctionEvaluator import *
